Models reogranization into whole clases
