+++ SafePlace/App.config
@@ -25,6 +25,5 @@
     <add key="azure-group-id" value="users" />
   </appSettings>
   <connectionStrings>
-  <!--  <add name="DataBase" providerName="System.Data.SqlClient" connectionString="Server=tcp:smartvision.database.windows.net,1433;Initial Catalog=SafePlaceDatabase;Persist Security Info=False;User ID=SmartVision2018;Password=TDVData2018;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30" /> -->
   <add name="DataContext" connectionString="Server=tcp:smartvision.database.windows.net,1433;Initial Catalog=SafePlaceDatabase;Persist Security Info=False;User ID=SmartVision2018;Password=TDVData2018;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;" providerName="System.Data.SqlClient" /></connectionStrings>
 </configuration>
