--- SafePlace/Models/DataBase/Floor.cs
@@ -1,29 +1,0 @@
-namespace SafePlace.Models
-{
-    using System;
-    using System.Collections.Generic;
-    using System.ComponentModel.DataAnnotations;
-    using System.ComponentModel.DataAnnotations.Schema;
-    using System.Data.Entity.Spatial;
-
-    [Table("Floor", Schema = "Topas")]
-    public partial class Floor : Model
-    {
-
-        public Floor()
-        {
-            this.Cameras = new HashSet<Camera>(); 
-        }
-
-        [Key]
-        public Guid Guid { get; set; }
-
-        [StringLength(255)]
-        public string ImagePath { get; set; }
-
-        [StringLength(64)]
-        public string Name { get; set; }
-        
-        public virtual ICollection<Camera> Cameras { get; set; }
-    }
-}
