+++ SafePlace/Models/Person.cs
@@ -1,24 +1,34 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace SafePlace.Models
 {
-    public partial class Person
+    using System;
+    using System.Collections.Generic;
+    using System.ComponentModel.DataAnnotations;
+    using System.ComponentModel.DataAnnotations.Schema;
+    using System.Data.Entity.Spatial;
+
+    [Table("Person", Schema = "Topas")]
+    public partial class Person : Model
     {
-        #region DB part contains
-        //public String Name { get; set; }
-        //public String LastName { get; set; }
-        //public ICollection<Guid> AllowedCameras { get; set; }
-        //public Guid Guid { get; set; }
+        #region database fields
+        [Key]
+        public Guid Guid { get; set; }
+
+        [Required]
+        [StringLength(64)]
+        public string Name { get; set; }
+
+        [Required]
+        [StringLength(64)]
+        public string LastName { get; set; }
+
         //Possible types: Employee (Various team roles + intern), Guest, BusinessCollaborator (visitors from foreign branches or related companies).
         //Could be in format: type, subtype. For example: "Visitor, business partner" or "Employee, intern".
-        //public String PersonType { get; set; }
+        public PersonType PersonType { get; set; }
+
+        public virtual ICollection<Camera> AllowedCameras { get; set; }
         #endregion
 
+        #region extr stuff
         //Whenever a camera notices a person, following should happen:
         //1. If camera was not null, the person is removed from the camera's SpottedPeople list;
         //2. A new value is set and the person is added there.
@@ -35,7 +45,6 @@ namespace SafePlace.Models
         {
             return $"{Name} {LastName}";
         }
-
-       
+        #endregion
     }
 }
