+++ SafePlace/SafePlace.csproj
@@ -162,17 +162,14 @@
     <Compile Include="Converters\SettingsModesToReadOnlyConverter.cs" />
     <Compile Include="Converters\StatusToColorConverter.cs" />
     <Compile Include="DBCommunication\DBCommunicator.cs" />
-    <Compile Include="Models\DataBase\Camera.cs" />
-    <Compile Include="Models\DataBase\Floor.cs" />
+    <Compile Include="Models\Camera.cs" />
+    <Compile Include="Models\Floor.cs" />
     <Compile Include="Models\Model.cs" />
-    <Compile Include="Models\DataBase\Person.cs" />
-    <Compile Include="Models\DataBase\PersonType.cs" />
+    <Compile Include="Models\Person.cs" />
+    <Compile Include="Models\PersonType.cs" />
     <Compile Include="DB\DataContext.cs" />
     <Compile Include="Enums\CameraStatus.cs" />
     <Compile Include="Enums\SettingsModes.cs" />
-    <Compile Include="Models\Dynamic\Camera.cs" />
-    <Compile Include="Models\Dynamic\Floor.cs" />
-    <Compile Include="Models\Dynamic\Person.cs" />
     <Compile Include="Service\CameraAnalyzeService.cs" />
     <Compile Include="Service\CameraService.cs" />
     <Compile Include="Service\ConsoleLogger.cs" />
